Most functions now moved from include/items.php
