+++ src/Protocol/Feed.php
@@ -423,7 +423,7 @@ class Feed {
 			if (!$simulate) {
 				logger("Stored feed: ".print_r($item, true), LOGGER_DEBUG);
 
-				$notify = item_is_remote_self($contact, $item);
+				$notify = Item::isRemoteSelf($contact, $item);
 
 				// Distributed items should have a well formatted URI.
 				// Additionally we have to avoid conflicts with identical URI between imported feeds and these items.
@@ -433,7 +433,7 @@ class Feed {
 					unset($item['parent-uri']);
 				}
 
-				$id = item_store($item, false, $notify);
+				$id = Item::insert($item, false, $notify);
 
 				logger("Feed for contact ".$contact["url"]." stored under id ".$id);
 			} else {
