+++ src/Worker/Expire.php
@@ -9,6 +9,7 @@ namespace Friendica\Worker;
 use Friendica\Core\Addon;
 use Friendica\Core\Config;
 use Friendica\Core\Worker;
+use Friendica\Model\Item;
 use Friendica\Database\DBM;
 use dba;
 
@@ -43,7 +44,7 @@ class Expire {
 			$user = dba::selectFirst('user', ['uid', 'username', 'expire'], ['uid' => $param]);
 			if (DBM::is_result($user)) {
 				logger('Expire items for user '.$user['uid'].' ('.$user['username'].') - interval: '.$user['expire'], LOGGER_DEBUG);
-				item_expire($user['uid'], $user['expire']);
+				Item::expire($user['uid'], $user['expire']);
 				logger('Expire items for user '.$user['uid'].' ('.$user['username'].') - done ', LOGGER_DEBUG);
 			}
 			return;
